requirements.txt
python-dotenv==1.1.1
streamlit==1.49.1
google-generativeai==0.8.0